+++ tests/test_audio_signal.py
@@ -542,6 +542,43 @@ class TestTrimFunction(unittest.TestCase):
         self.assertTrue(np.allclose(result[:, 1], np.arange(1, 5)))
 
 
+class TestEnergyFunction(unittest.TestCase):
+
+    def test_types(self):
+        # mono signals
+        result = energy(sig_1d)
+        self.assertIsInstance(result, float)
+        # multi-channel signals
+        result = energy(sig_2d)
+        self.assertIsInstance(result, float)
+
+    def test_values(self):
+        # mono signals
+        result = energy(sig_1d)
+        self.assertTrue(np.allclose(result, 3))
+        result = energy(np.zeros(100))
+        self.assertTrue(np.allclose(result, 0))
+        # multi-channel signals
+        result = energy(sig_2d)
+        self.assertTrue(np.allclose(result, 8))
+        result = energy(np.zeros(100).reshape(-1, 2))
+        self.assertTrue(np.allclose(result, 0))
+
+    def test_frames(self):
+        # mono signals
+        frames = FramedSignal(sig_1d, frame_size=4, hop_size=2)
+        result = energy(frames)
+        self.assertTrue(np.allclose(result, [0, 1, 2, 1, 1]))
+        result = energy(np.zeros(100))
+        self.assertTrue(np.allclose(result, 0))
+        # multi-channel signals
+        frames = FramedSignal(sig_2d, frame_size=4, hop_size=2)
+        result = energy(frames)
+        self.assertTrue(np.allclose(result, [1, 3, 4, 3, 3]))
+        result = energy(np.zeros(100).reshape(-1, 2))
+        self.assertTrue(np.allclose(result, 0))
+
+
 class TestRootMeanSquareFunction(unittest.TestCase):
 
     def test_types(self):
@@ -561,6 +598,22 @@ class TestRootMeanSquareFunction(unittest.TestCase):
         # multi-channel signals
         result = root_mean_square(sig_2d)
         self.assertTrue(np.allclose(result, 2. / 3))
+        result = root_mean_square(np.zeros(100).reshape(-1, 2))
+        self.assertTrue(np.allclose(result, 0))
+
+    def test_frames(self):
+        # mono signals
+        frames = FramedSignal(sig_1d, frame_size=4, hop_size=2)
+        result = root_mean_square(frames)
+        self.assertTrue(np.allclose(result, [0, 0.5, 0.70710678, 0.5, 0.5]))
+        result = root_mean_square(np.zeros(100))
+        self.assertTrue(np.allclose(result, 0))
+        # multi-channel signals
+        frames = FramedSignal(sig_2d, frame_size=4, hop_size=2)
+        result = root_mean_square(frames)
+        self.assertTrue(np.allclose(result, [0.35355339, 0.61237244,
+                                             0.70710678, 0.61237244,
+                                             0.61237244]))
         result = root_mean_square(np.zeros(100).reshape(-1, 2))
         self.assertTrue(np.allclose(result, 0))
 
@@ -605,6 +658,23 @@ class TestSoundPressureLevelFunction(unittest.TestCase):
         sig = remix(sinus_int16, 2)
         result = sound_pressure_level(sig)
         self.assertTrue(np.allclose(result, 0.))
+
+    def test_frames(self):
+        # mono signals
+        frames = FramedSignal(sig_1d, frame_size=4, hop_size=2)
+        result = sound_pressure_level(frames)
+        self.assertTrue(np.allclose(result, [-np.finfo(float).max, -6.0206,
+                                             -3.0103, -6.0206, -6.0206]))
+        result = sound_pressure_level(np.zeros(100))
+        self.assertTrue(np.allclose(result, -np.finfo(float).max))
+        # multi-channel signals
+        frames = FramedSignal(sig_2d, frame_size=4, hop_size=2)
+        result = sound_pressure_level(frames)
+        self.assertTrue(np.allclose(result, [-9.03089987, -4.25968732,
+                                             -3.01029996, -4.25968732,
+                                             -4.25968732]))
+        result = sound_pressure_level(np.zeros(100).reshape(-1, 2))
+        self.assertTrue(np.allclose(result, -np.finfo(float).max))
 
 
 class TestLoadWaveFileFunction(unittest.TestCase):
@@ -936,6 +1006,19 @@ class TestSignalClass(unittest.TestCase):
         orig.write(tmp_file)
         result = Signal(tmp_file)
         self.assertTrue(np.allclose(orig, result))
+
+    def test_methods(self):
+        # mono signals
+        signal = Signal(sig_1d)
+        self.assertTrue(np.allclose(signal.energy(), 3))
+        self.assertTrue(np.allclose(signal.rms(), 0.57735026919))
+        self.assertTrue(np.allclose(signal.spl(), -4.7712125472))
+        # multi-channel signals
+        signal = Signal(sig_2d)
+        self.assertTrue(np.allclose(signal.energy(), 8))
+        self.assertTrue(np.allclose(signal.root_mean_square(), 2. / 3))
+        self.assertTrue(np.allclose(signal.sound_pressure_level(),
+                                    -3.52182518111))
 
 
 class TestSignalProcessorClass(unittest.TestCase):
@@ -1387,6 +1470,25 @@ class TestFramedSignalClass(unittest.TestCase):
         result = FramedSignal(sample_file, fps=50)
         self.assertTrue(result.frame_size == 2048)
         self.assertTrue(result.hop_size == 882.)
+
+    def test_methods(self):
+        # mono signals
+        frames = FramedSignal(sig_1d, frame_size=4, hop_size=2)
+        self.assertTrue(np.allclose(frames.energy(), [0, 1, 2, 1, 1]))
+        self.assertTrue(np.allclose(frames.rms(),
+                                    [0, 0.5, 0.70710678, 0.5, 0.5]))
+        self.assertTrue(np.allclose(frames.spl(),
+                                    [-np.finfo(float).max, -6.0206, -3.0103,
+                                     -6.0206, -6.0206]))
+        # multi-channel signals
+        frames = FramedSignal(sig_2d, frame_size=4, hop_size=2)
+        self.assertTrue(np.allclose(frames.energy(), [1, 3, 4, 3, 3]))
+        self.assertTrue(np.allclose(frames.root_mean_square(),
+                                    [0.35355339, 0.61237244, 0.70710678,
+                                     0.61237244, 0.61237244]))
+        self.assertTrue(np.allclose(frames.sound_pressure_level(),
+                                    [-9.03089987, -4.25968732, -3.01029996,
+                                     -4.25968732, -4.25968732]))
 
 
 class TestFramedSignalProcessorClass(unittest.TestCase):